feat(deps)!: make aws-sdk a peer/dev dependency
This partially addresses the problem raised in issue #29. It allows for
consumers to reliably load only their own AWS SDK version rather than
risking potentially bundling multiple copies in a build. It also makes
this package more Lambda function friendly, since the aws-sdk version is
vended by the runtime by default and does not need to be packaged in.

Note that this change does not make this package agnostic to the choice
between AWS SDK v2 and v3. The current implementation still relies on
imports that exist on v2 but not v3.